Метод подготовки данных вынес в интерфейс и реализовал внедрение зависимости
